% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hello.R
\name{hello}
\alias{hello}
\title{Hello!}
\usage{
hello(name = "world")
}
\arguments{
\item{name}{Character to greet. Defaults to \code{world}.}
}
\value{
Character greeting.
}
\description{
Trivial example for documentation.
}
\examples{
hello()
hello("Stan")
}
