% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_pkgdown_gha.R
\name{use_pkgdown_gha}
\alias{use_pkgdown_gha}
\title{Copy the Stan pkgdown GitHub Action}
\usage{
use_pkgdown_gha(pkg = ".")
}
\arguments{
\item{pkg}{Path to the package.}
}
\value{
The path of the workflow, invisibly.
}
\description{
Copies this package's \code{pkgdown.yaml} workflow into \code{.github/workflows}.
}
